Use name attr instead of title
Based on what I'm seeing in [the proposed solution code here](https://github.com/joinpursuit/FSW-React-Forms-Exercise/blob/forms-exercise-2/src/UserInfoForm.js#L31), it looks like we want to be going with the `name` attribute on the inputs rather than `title`. If this is not the case, then this PR can be closed and we'll need to make the update to the solution repo.